nested bhomObject serialised via Json. Tidy up. TODO: Pull
