+++ Speckle_Adapter/AdapterActions/Pull.cs
@@ -7,62 +7,38 @@ using System.Threading.Tasks;
 using BH.oM.Base;
 using BH.oM.Data.Requests;
 using SpeckleCore;
+using BH.Engine.Speckle;
 
 namespace BH.Adapter.Speckle
 {
     public partial class SpeckleAdapter : BHoMAdapter
     {
-        public override IEnumerable<object> Pull(IRequest query, Dictionary<string, object> config = null)
+        public override IEnumerable<object> Pull(IRequest request, Dictionary<string, object> config = null)
         {
-            var response = SpeckleClient.StreamGetObjectsAsync(SpeckleStreamId, "").Result;
+            if (request == null)
+                return ReadAll();
+
+            // Make sure this is a FilterRequest
+            FilterRequest filterRequest = request as FilterRequest;
+            if (filterRequest == null)
+            {
+                Engine.Reflection.Compute.RecordWarning("Please specify a FilterRequest");
+                return new List<object>();
+            }
 
             List<IBHoMObject> bHoMObjects = new List<IBHoMObject>();
             List<IObject> iObjects = new List<IObject>();
             List<object> reminder = new List<object>();
 
-            bool assignSpeckleIdToBHoMObjects = true;
+            Read(filterRequest, out bHoMObjects, out iObjects, out reminder);
 
-            if (query == null)
-            {
-
-                if (!BH.Engine.Speckle.Convert.ToBHoM(response, out bHoMObjects, out iObjects, out reminder, assignSpeckleIdToBHoMObjects))
-                    BH.Engine.Reflection.Compute.RecordError("Failed to deserialize and cast the Server response into BHoM objects.");
-
-                return bHoMObjects.Concat(iObjects).Concat(reminder);
-            }
+            // Return stuff
+            if (typeof(IBHoMObject).IsAssignableFrom(filterRequest.Type))
+                return bHoMObjects;
+            else if (typeof(IObject).IsAssignableFrom(filterRequest.Type))
+                return iObjects;
             else
-            {
-                /// -------------------
-                /// Base Pull rewritten
-                /// -------------------
-
-                // Make sure this is a FilterQuery
-                FilterRequest filter = query as FilterRequest;
-                if (filter == null)
-                {
-                    Engine.Reflection.Compute.RecordWarning("Please specify a FilterQuery");
-                    return new List<object>();
-                }
-
-                List<string> speckleIds = QueryToSpeckleIds(filter);
-
-                // Read the IBHoMObjects
-                BH.Engine.Speckle.Convert.ToBHoM(response, out bHoMObjects, out iObjects, out reminder, assignSpeckleIdToBHoMObjects, speckleIds);
-
-                // Filter by tag if any 
-                bHoMObjects = filter.Tag == "" ? bHoMObjects : bHoMObjects.Where(x => x.Tags.Contains(filter.Tag)).ToList();
-
-                // Return stuff
-                if (typeof(IBHoMObject).IsAssignableFrom(filter.Type))
-                    return bHoMObjects;
-                else if (typeof(IObject).IsAssignableFrom(filter.Type))
-                    return iObjects;
-                else
-                    return bHoMObjects.Concat(iObjects).Concat(reminder);
-
-            }
-
-            return new List<object>();
+                return bHoMObjects.Concat(iObjects).Concat(reminder);
         }
     }
 }
