+++ Speckle_Adapter/CRUD/Read.cs
@@ -6,123 +6,49 @@ using System.Reflection;
 using BH.oM.Base;
 using BH.oM.Data.Requests;
 using SpeckleCore;
+using BH.Engine.Speckle;
+
 
 namespace BH.Adapter.Speckle
 {
     public partial class SpeckleAdapter
     {
-        /***************************************************/
-        /**** Adapter overload method                   ****/
-        /***************************************************/
         protected override IEnumerable<IBHoMObject> Read(Type type, IList ids)
         {
-            // // - This method will only be called from the BHoM_Adapter by either: 1) The Replace method 2) The Read overload with the query stuff in it
+            // Not used. Override required due to `abstract` in base adapter. To be removed after Refactoring Level 04.
+            return null;
+        }
 
-            ////Choose what to pull out depending on the type. Also see example methods below for pulling out bars and dependencies
-            //if (type == typeof(Node))
-            //    return ReadNodes(ids as dynamic);
-            //else if (type == typeof(Bar))
-            //    return ReadBars(ids as dynamic);
-            //else if (type == typeof(ISectionProperty) || type.GetInterfaces().Contains(typeof(ISectionProperty)))
-            //    return ReadSectionProperties(ids as dynamic);
-            //else if (type == typeof(Material))
-            //    return ReadMaterials(ids as dynamic);
 
-            ///------
-            ///
+        protected IEnumerable<object> ReadAll(bool assignSpeckleIdToBHoMObjects = true)
+        {
+            var response = SpeckleClient.StreamGetObjectsAsync(SpeckleStreamId, "").Result;
 
-            List<string> speckleGuids = new List<string>();
-
-            for (int i = 0; i < ids?.Count; i++)
-                speckleGuids.Add(ids[i].ToString());
-
-            ResponseObject response = MakeGETRequest(speckleGuids);               
-            
-            List <IBHoMObject> bHoMObjects = new List<IBHoMObject>();
+            List<IBHoMObject> bHoMObjects = new List<IBHoMObject>();
             List<IObject> iObjects = new List<IObject>();
             List<object> reminder = new List<object>();
 
-            bool assignSpeckleIdToBHoMObjects = true;
-
             if (!BH.Engine.Speckle.Convert.ToBHoM(response, out bHoMObjects, out iObjects, out reminder, assignSpeckleIdToBHoMObjects))
-            {
-                Engine.Reflection.Compute.RecordError("Failed to elaborate server response.");
-                return new List<IBHoMObject>();
-            }
+                BH.Engine.Reflection.Compute.RecordError("Failed to deserialize and cast the Server response into BHoM objects.");
 
-            return bHoMObjects;
+            return bHoMObjects.Concat(iObjects).Concat(reminder);
         }
 
-  
-        public ResponseObject MakeGETRequest(List<string> speckleGuids = null)
+        protected bool Read(FilterRequest filterRequest, out List<IBHoMObject> bHoMObjects, out List<IObject> iObjects, out List<object> reminder)
         {
-            if (speckleGuids == null)
-            {
-                return SpeckleClient.StreamGetObjectsAsync(SpeckleStreamId, "").Result;      
-            }
-            else
-            {
-                // GET only specific IDs
-                return SpeckleClient.ObjectGetBulkAsync(speckleGuids.ToArray(), "omit=displayValue").Result;
-            }
+            var response = SpeckleClient.StreamGetObjectsAsync(SpeckleStreamId, "").Result;
+
+            // Extract the speckleIds for selection from the FilterRequest.
+            List<string> speckleIds = Query.SpeckleIds(filterRequest);
+            speckleIds = speckleIds?.Count != 0 ? speckleIds : null;
+
+            // Convert the response to the appropriate object types.
+            BH.Engine.Speckle.Convert.ToBHoM(response, out bHoMObjects, out iObjects, out reminder, speckleIds?.Count != 0, speckleIds);
+
+            // Filter by tag if any 
+            bHoMObjects = filterRequest.Tag == "" ? bHoMObjects : bHoMObjects.Where(x => x.Tags.Contains(filterRequest.Tag)).ToList();
+
+            return true;
         }
-
-        public List<string> QueryToSpeckleIds(FilterRequest query)
-        {
-            List<string> speckleGuids = new List<string>();
-
-            IList objectIds = null;
-            object idObject;
-            if (query.Equalities.TryGetValue("ObjectIds", out idObject) && idObject is IList)
-                objectIds = idObject as IList;
-
-            if (objectIds == null)
-                return speckleGuids;
-
-            for (int i = 0; i < objectIds.Count; i++)
-                speckleGuids.Add(objectIds[i].ToString());
-
-            return speckleGuids;
-        }
-
-        /***************************************************/
-        /**** Private specific read methods             ****/
-        ///***************************************************/
-
-        ////The List<string> in the methods below can be changed to a list of any type of identification more suitable for the toolkit
-
-        //private List<Bar> ReadBars( List<string> ids = null )
-        //{
-        //  //Implement code for reading bars
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<Node> ReadNodes( List<string> ids = null )
-        //{
-        //  //Implement code for reading nodes
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<ISectionProperty> ReadSectionProperties( List<string> ids = null )
-        //{
-        //  //Implement code for reading section properties
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<Material> ReadMaterials( List<string> ids = null )
-        //{
-        //  //Implement code for reading materials
-        //  throw new NotImplementedException();
-        //}
-
-        /***************************************************/
-
-
     }
 }
