+++ Speckle_Adapter/Speckle_Adapter.csproj
@@ -140,6 +140,8 @@
   <ItemGroup>
     <Compile Include="AdapterActions\Pull.cs" />
     <Compile Include="AdapterActions\Push.cs" />
+    <Compile Include="AdapterActions\_PushMethods\BasePushRewritten.cs" />
+    <Compile Include="AdapterActions\_PushMethods\ConfigureHistory.cs" />
     <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\Delete.cs" />
     <Compile Include="CRUD\Read.cs" />
