+++ Speckle_Engine/Convert/ToBHoM.cs
@@ -37,21 +37,21 @@ namespace BH.Engine.Speckle
         public static IEnumerable<IBHoMObject> ToBHoM(ResponseObject response, bool setAssignedId = true, List<string> speckleIds = null)
         {
             List<IBHoMObject> bhomObjects = new List<IBHoMObject>();
-
             List<IObject> iObjects;
             List<object> reminder;
+
             ToBHoM(response, out bhomObjects, out iObjects, out reminder, setAssignedId, speckleIds);
 
             return bhomObjects;
         }
 
-        public static bool ToBHoM(ResponseObject response, out List<IBHoMObject> bHoMObjects, out List<IObject> iObjects, out List<object> reminder, bool assignSpeckleIdToBHoMObjects = true, List<string> speckleIds = null)
+        public static bool ToBHoM(ResponseObject response, 
+                                    out List<IBHoMObject> bHoMObjects, out List<IObject> iObjects, out List<object> reminder, 
+                                    bool assignSpeckleIdToBHoMObjects = true, List<string> speckleIds = null)
         {
             bHoMObjects = new List<IBHoMObject>();
             iObjects = new List<IObject>();
             reminder = new List<object>();
-
-
 
             for (int i = 0; i < response.Resources.Count; i++)
             {
@@ -59,11 +59,9 @@ namespace BH.Engine.Speckle
                 var gasd = Converter.Deserialise(response.Resources[i].Properties.Values.OfType<SpeckleObject>());
                 var nestedBHoMProperties = response.Resources[i].Properties;
 
-
-           
-
+                // If we are filtering by speckleId, skip the object if its speckleId doesn't match.
                 if (speckleIds != null && speckleIds.Count > 0)
-                    if (!speckleIds.Any(id => id == response.Resources[i]._id))
+                    if (!speckleIds.Any(id => id == response.Resources[i]._id)) // note: slow, o(n²)
                         continue;
 
                 IBHoMObject iBHoMObject = resource as IBHoMObject;
