+++ Speckle_Engine/packages.config
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Countly" version="19.8.0" targetFramework="net452" />
   <package id="DeviceId" version="4.3.0" targetFramework="net452" />
@@ -6,6 +6,7 @@
   <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net452" />
   <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net452" />
+  <package id="RhinoCommon" version="5.12.50810.13095" targetFramework="net461" />
   <package id="SpeckleCore" version="1.6.9" targetFramework="net452" />
   <package id="SpeckleCoreGeometry" version="1.2.8" targetFramework="net461" />
   <package id="sqlite-net-pcl" version="1.5.231" targetFramework="net452" />
